data_lake/requirements.txt
faker==20.1.0
numpy==1.26.2
pandas==2.1.4
jsonschema==4.20.0
python-dateutil==2.8.2
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.5
python-socketio==5.10.0
eventlet==0.33.3
python-dotenv==1.0.0
requests==2.31.0
scikit-learn==1.3.2
joblib==1.3.2
